refactor(#1931): decompose detectEarlyErrors; run on multi-source; drop dead treeshake option
Split the ~3,350-line detectEarlyErrors monolith in validation.ts into
per-concern modules under src/compiler/early-errors/, sharing one AST walk
through an EarlyErrorContext:

- context.ts        EarlyErrorContext + factory
- predicates.ts     34 pure, stateless predicate/collection helpers
- assignment.ts     AssignmentPattern validators
- duplicates.ts     duplicate params / lexical / private-name / switch rules
- labels.ts         duplicate-label rules
- tdz.ts            temporal-dead-zone reference checks
- node-checks.ts    per-node dispatch (former `visit` walk)
- module-rules.ts   export-default / dup-export-names / module-item position /
                    reserved yield-await / HTML close comment / dup constructors
- index.ts          thin orchestrator (the new detectEarlyErrors)

validation.ts: 3,647 -> 283 lines. detectEarlyErrors is re-exported from
validation.ts so existing import sites (compiler.ts) are unchanged.

Behaviour-preserving: a 447-case oracle (snippets per rule + 400 test262
negative.phase:parse files) is byte-identical before/after.

Run on every path: compileMultiSource and compileFilesSource now run
detectEarlyErrors on every user source file (skipping allowJs deps), so a
duplicate-`let` is rejected in multi-file compiles too. The single-source
compile() path used by test262 is unchanged.

treeshake: removed the dead CompileOptions.treeshake flag (never read by any
compile path). The standalone treeshake() helper stays exported and tested.

Also fixed the stale isStrictMode doc comment claiming modules are strict.

Tests: tests/issue-1931.test.ts — 20 unit/integration tests covering
predicates, rule modules, full detectEarlyErrors, and the multi-source
duplicate-`let` rejection.

✓
