+++ src/compiler.ts
@@ -1046,6 +1046,32 @@ export async function compileMultiSource(
     };
   }
 
+  // Early error detection — catch ES-spec syntax errors that TypeScript misses,
+  // on every user source file (#1931). Previously the multi-source path skipped
+  // ES early errors entirely; now compileSource and compileMultiSource share the
+  // same detectEarlyErrors pass so e.g. a duplicate-`let` is rejected in a
+  // multi-file compile too. allowJs dependency files are skipped (their JS may
+  // use patterns we cannot control) — same scoping as the diagnostic loop above.
+  if (!options.allowJs) {
+    for (const sf of multiAst.sourceFiles) {
+      errors.push(...detectEarlyErrors(sf));
+    }
+    if (errors.some((e) => e.severity === "error")) {
+      return {
+        binary: new Uint8Array(0),
+        wat: "",
+        dts: "",
+        importsHelper: "",
+        success: false,
+        errors,
+        stringPool: [],
+        imports: [],
+        hasMain: false,
+        hasTopLevelStatements: false,
+      };
+    }
+  }
+
   // Safe mode validation for all source files
   if (options.safe) {
     for (const sf of multiAst.sourceFiles) {
@@ -1330,6 +1356,29 @@ export async function compileFilesSource(entryPath: string, options: CompileOpti
       hasMain: false,
       hasTopLevelStatements: false,
     };
+  }
+
+  // Early error detection — catch ES-spec syntax errors that TypeScript misses,
+  // on every user source file (#1931). compileFilesSource previously skipped ES
+  // early errors; wire the same detectEarlyErrors pass here too.
+  if (!options.allowJs) {
+    for (const sf of multiAst.sourceFiles) {
+      errors.push(...detectEarlyErrors(sf));
+    }
+    if (errors.some((e) => e.severity === "error")) {
+      return {
+        binary: new Uint8Array(0),
+        wat: "",
+        dts: "",
+        importsHelper: "",
+        success: false,
+        errors,
+        stringPool: [],
+        imports: [],
+        hasMain: false,
+        hasTopLevelStatements: false,
+      };
+    }
   }
 
   // Safe mode validation for all source files
